Emerald

Types

Environment

Emerald.AbstractSoilVC — Type

abstract type AbstractSoilVC{FT<:AbstractFloat}

Hierachy of AbstractSoilVC:

BrooksCorey

VanGenuchten

Emerald.BrooksCorey — Type

struct BrooksCorey{FT<:AbstractFloat} <: Emerald.AbstractSoilVC{FT<:AbstractFloat}

Brooks Corey soil parameters

Fields

TYPE::String

Soil type

B::AbstractFloat

Soil b

Ψ_SAT::AbstractFloat

Potential at saturation [MPa]

Θ_SAT::AbstractFloat

Saturated soil volumetric water content

Θ_RES::AbstractFloat

Residual soil volumetric water content

Examples

bc = BrooksCorey{FT}("Test", FT(5), FT(2), FT(0.5), FT(0.1));
bc = BrooksCorey{FT}(VanGenuchten{FT}("Loam"));

Emerald.VanGenuchten — Type

struct VanGenuchten{FT<:AbstractFloat} <: Emerald.AbstractSoilVC{FT<:AbstractFloat}

van Genuchten soil parameters

Fields

TYPE::String

Soil type

Α::AbstractFloat

Soil α is related to the inverse of the air entry suction, α > 0

N::AbstractFloat

Soil n is Measure of the pore-size distribution

M::AbstractFloat

Soil m = 1 - 1/n

Θ_SAT::AbstractFloat

Saturated soil volumetric water content

Θ_RES::AbstractFloat

Residual soil volumetric water content

Examples

vg = VanGenuchten{FT}("Loam");
vg = VanGenuchten{FT}("Silt");
vg = VanGenuchten{FT}("Test", FT(100), FT(2), FT(0.5), FT(0.1));

Emerald.SoilAir — Type

mutable struct SoilAir{FT<:AbstractFloat}

Struct that contains environmental conditions, such as soil moisture and atmospheric vapor pressure. Note that this structure is designed to be containers to interact with other CliMA modules and to prescribe values.

Fields

AREA::AbstractFloat

Total area of the soil/air interface [m²]

COLOR::Int64

Soil color class used for soil albedo calculations

N_AIR::Int64

Number of air layers

N_SOIL::Int64

Number of soil layers

VC::Union{Array{BrooksCorey{FT}, 1}, Array{VanGenuchten{FT}, 1}} where FT<:AbstractFloat

Soil moisture retention curve

Z_AIR::Vector{FT} where FT<:AbstractFloat

Z profile for air [m]

Z_SOIL::Vector{FT} where FT<:AbstractFloat

Z profile for soil [m]

ΔZ_AIR::Vector{FT} where FT<:AbstractFloat

ΔZ profile for air [m]

ΔZ_SOIL::Vector{FT} where FT<:AbstractFloat

ΔZ profile for soil [m]

p_CO₂::Vector{FT} where FT<:AbstractFloat

CO₂ partial pressure at different air layers [Pa]

p_H₂O::Vector{FT} where FT<:AbstractFloat

H₂O partial pressure at different air layers [Pa]

t_air::Vector{FT} where FT<:AbstractFloat

Temperature at different air layers [K]

t_soil::Vector{FT} where FT<:AbstractFloat

Temperature at different soil layers [K]

wind::Vector{FT} where FT<:AbstractFloat

Wind speed (total) [m s⁻¹]

wind_z::Vector{FT} where FT<:AbstractFloat

Wind speed (vertical) [m s⁻¹]

θ::Vector{FT} where FT<:AbstractFloat

Soil water content at different soil layers [m³ m⁻³]

p_H₂O_sat::Vector{FT} where FT<:AbstractFloat

Saturated vapor pressure at different air layers [Pa]

vpd::Vector{FT} where FT<:AbstractFloat

Vapor pressure deficit at different air layers [Pa]

ψ::Vector{FT} where FT<:AbstractFloat

Soil water potential at different soil layers [MPa]

Examples

;

Leaf

Emerald.Leaf — Type

mutable struct Leaf{FT<:AbstractFloat}

Struct that contains leaf traits such as plant photosynthesis, hydraulics, and biophysics.

Fields

BIO_PHYSICS::Emerald.LeafBiophysics

Biophysical parameter structure

t::AbstractFloat

Current leaf temperature

_t::AbstractFloat

Last leaf temperature. If different from t, then make temperature correction

Examples

leaf = Leaf{FT}();
leaf = Leaf{FT}(collect(FT,400:5:2500.1));
leaf = Leaf{FT}(collect(FT,400:5:2500.1); opti=Emerald.OPTI_2021);
leaf = Leaf{FT}(WaveLengthSet{FT}());
leaf = Leaf{FT}(WaveLengthSet{FT}(); opti=Emerald.OPTI_2021);

Emerald.LeafBiophysics — Type

mutable struct LeafBiophysics{FT<:AbstractFloat}

Struct that contains leaf biophysical traits used to run leaf reflection and transmittance.

Fields

FQE::AbstractFloat

Leaf fluorescence quantum efficiency (Fo standard)

LMA::AbstractFloat

Dry matter content (dry leaf mass per unit area) [g cm⁻²]

CBC::AbstractFloat

Carbon-based constituents in LMA [g cm⁻²]

PRO::AbstractFloat

Protein content in LMA (PRO = LMA - CBC) [g cm⁻²]

K_ANT::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of anthocynanin [-]

K_CAB::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of chlorophyll a and b [-]

K_CAR_V::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of violaxanthin carotenoid [-]

K_CAR_Z::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of zeaxanthin carotenoid [-]

K_CBC::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of carbon-based constituents [-]

K_H₂O::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of water [-]

K_LMA::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of dry matter [-]

K_PRO::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of protein [-]

K_PS::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of PS I and II [-]

K_SENES::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of senescent material [-]

MESOPHYLL_N::AbstractFloat

Leaf mesophyll structural parameter that describes mesophyll reflection and transmittance

NDUB::Int64

Doubling adding layers

NR::Vector{FT} where FT<:AbstractFloat

Refractive index [-]

ant::AbstractFloat

Anthocynanin content [ug cm⁻²]

cab::AbstractFloat

Chlorophyll a and b content [ug cm⁻²]

car::AbstractFloat

Carotenoid content [ug cm⁻²]

f_senes::AbstractFloat

Senescent material (brown pigments) fraction [-]

f_zeax::Any

Zeaxanthin fraction in Carotenoid (1=all Zeaxanthin, 0=all Violaxanthin) [-]

l_H₂O::AbstractFloat

Equivalent water thickness [cm]

mat_b::Matrix{FT} where FT<:AbstractFloat

Fluorescence excitation matrix backwards [-]

mat_f::Matrix{FT} where FT<:AbstractFloat

Fluorescence excitation matrix forwards [-]

α_cab::Vector{FT} where FT<:AbstractFloat

Relative absorption by Chlorophyll [-]

α_cabcar::Vector{FT} where FT<:AbstractFloat

Relative absorption by Chlorophyll+Carotenoid [-]

ρ_LW::AbstractFloat

Broadband thermal reflectance, related to blackbody emittance [-]

ρ_SW::Vector{FT} where FT<:AbstractFloat

Shortwave leaf reflectance [-]

τ_LW::AbstractFloat

Broadband thermal transmission, related to blackbody emittance [-]

τ_SW::Vector{FT} where FT<:AbstractFloat

Shortwave leaf transmission [-]

_k_all::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of all materials

_α_SW::Vector{FT} where FT<:AbstractFloat

Shortwave absorption, 1 .- ρSW .- τSW [-]

Examples

lbio = Emerald.LeafBiophysics{FT}();
lbio = Emerald.LeafBiophysics{FT}(collect(FT,400:5:2500.1));
lbio = Emerald.LeafBiophysics{FT}(collect(FT,400:5:2500.1); opti=Emerald.OPTI_2021);
lbio = Emerald.LeafBiophysics{FT}(WaveLengthSet{FT}());
lbio = Emerald.LeafBiophysics{FT}(WaveLengthSet{FT}(); opti=Emerald.OPTI_2021);

Radiation

Emerald.WaveLengthSet — Type

mutable struct WaveLengthSet{FT<:AbstractFloat}

Structure that stores wave length information.

IΛ_NIR::Vector{Int64}

Indicies of Λ_NIR in Λ

IΛ_PAR::Vector{Int64}

Indicies of Λ_PAR in Λ

IΛ_SIF::Vector{Int64}

Indicies of Λ_SIF in Λ

IΛ_SIFE::Vector{Int64}

Indicies of Λ_SIFE in Λ

NΛ::Int64

Number of wavelength bins

NΛ_PAR::Int64

Number of wavelength bins for PAR

NΛ_SIF::Int64

Number of wavelength bins for SIF

NΛ_SIFE::Int64

Number of wavelength bins for SIF excitation

WL_NIR::Vector{FT} where FT<:AbstractFloat

Wavelength limits for NIR [nm]

WL_PAR::Vector{FT} where FT<:AbstractFloat

Wavelength limits for PAR [nm]

WL_SIF::Vector{FT} where FT<:AbstractFloat

Wavelength limits for SIF emission [nm]

WL_SIFE::Vector{FT} where FT<:AbstractFloat

Wavelength limits for SIF excitation [nm]

SΛ::Vector{FT} where FT<:AbstractFloat

Standard wavelength (boundaries) [nm]

ΔΛ::Vector{FT} where FT<:AbstractFloat

Differential wavelength [nm]

ΔΛ_PAR::Vector{FT} where FT<:AbstractFloat

Differential wavelength for PAR [nm]

ΔΛ_SIFE::Vector{FT} where FT<:AbstractFloat

Differential wavelength for SIF excitation [nm]

Λ::Vector{FT} where FT<:AbstractFloat

Wavelength (bins) [nm]

Λ_PAR::Vector{FT} where FT<:AbstractFloat

Wavelength bins for PAR [nm]

Λ_SIF::Vector{FT} where FT<:AbstractFloat

Wavelength bins for SIF [nm]

Λ_SIFE::Vector{FT} where FT<:AbstractFloat

Wavelength bins for SIF excitation [nm]

Examples

wls = WaveLengthSet{FT}();
wls = WaveLengthSet{FT}(collect(FT,400:5:2500));
wls = WaveLengthSet{FT}(collect(FT,400:5:2500); opti=Emerald.OPTI_2017);

Folium

Soli

Emerald.Soli.soil_ψ_25 — Function

This function calculates soil metric potential from soil retention curve type and soil volumetric water potential. The supported methods are

soil_ψ_25(vc, θ)

defined at /home/runner/work/Emerald.jl/Emerald.jl/src/Soli.jl:76.

soil_ψ_25(vc, θ)

defined at /home/runner/work/Emerald.jl/Emerald.jl/src/Soli.jl:91.

Emerald.Soli.soil_ψ_25 — Method

soil_ψ_25(vc::Union{BrooksCorey{FT}, VanGenuchten{FT}}, θ::FT) where {FT<:AbstractFloat}

Return the soil metric potential, given

vc BrooksCorey or VanGenuchten type structure

θ Soil volumetric water content (absolute value)

Examples

soil_ψ_25(bc, FT(0.2))
soil_ψ_25(vg, FT(0.2))

Utilities

Emerald.calctav — Function

calctav(α::FT, nr::FT) where {FT<:AbstractFloat}

Computes transmission of isotropic radiation across an interface between two dielectrics (Stern F., 1964; Allen W.A., 1973)). From calctav.m in PROSPECT-D

α angle of incidence

nr Index of refraction

Emerald.nanmean — Function

nanmean(x::Array)

Return the mean of array by ommiting the NaN, given

x Array of numbers, can be NaN

Example

xs = [1, 2, 4, NaN];
nmean = nanmean(xs);

Emerald.transmittance — Function

transmittance(θ_in::FT, n::FT) where {FT <:AbstractFloat}

Transmittance of S and P polarizations for a plane dielectric surface for light incident, given

θ_i Incoming light angle

n Relative index of refraction